% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoMCPcounter.R
\docType{data}
\name{annoMCPcounter}
\alias{annoMCPcounter}
\title{MCPcounter}
\format{
A list object.
}
\usage{
annoMCPcounter
}
\description{
probesets = read.table(curl('http://raw.githubusercontent.com/ebecht/MCPcounter/master/Signatures/probesets.txt'),
sep='\t',stringsAsFactors=FALSE,colClasses='character')
genes = read.table(curl('http://raw.githubusercontent.com/ebecht/MCPcounter/master/Signatures/genes.txt'),
                   sep='\t',
                   stringsAsFactors=FALSE,
                   header=TRUE,
                   colClasses='character',
                   check.names=FALSE)
weChat:MedBioInfoCloud
}
\author{
DoubleHelix
}
\keyword{datasets}
