% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsEffector.R
\docType{data}
\name{gsEffector}
\alias{gsEffector}
\title{gsEffector
Classification or annotation of some genes for analysis.
weChat:MedBioInfoCloud}
\format{
A list object.
}
\usage{
gsEffector
}
\description{
gsEffector
Classification or annotation of some genes for analysis.
weChat:MedBioInfoCloud
}
\author{
DoubleHelix
}
\keyword{datasets}
