% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsGSVA.R
\docType{data}
\name{gsGSVA}
\alias{gsGSVA}
\title{gsGSVA}
\format{
A list object.
}
\usage{
gsGSVA
}
\description{
The getGmt function reads data from the GMT file.
weChat:MedBioInfoCloud
}
\author{
DoubleHelix
}
\keyword{datasets}
