% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsMSigDB.R
\docType{data}
\name{gsMSigDB}
\alias{gsMSigDB}
\title{msigdb}
\format{
A GeneSet object.
}
\usage{
gsMSigDB
}
\description{
http://www.gsea-msigdb.org/gsea/index.jsp
weChat:MedBioInfoCloud
}
\author{
DoubleHelix
}
\keyword{datasets}
