% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMSigDB.R
\name{listMSigDB}
\alias{listMSigDB}
\title{View the dataset filename for MSigDB}
\usage{
listMSigDB()
}
\value{
A vector
}
\description{
View the dataset filename for MSigDB.
}
\examples{
listMSigDB()
}
\author{
DoubleHelix
}
